requirements.txt
PyYAML>=6.0.1
schema>=0.7.7